build: Remove coveralls as a dpendency.
We don't use coveralls to aggregate and verify coverage anymore.  We use
codecove as a part of CI but we don't need it as a python build
dependency anymore with the way we use it in github actions.  We still
install coverage locally to be able to generate the coverage data tha
codcov looks for.