test: Update pylint supressions.
Remove some previous pylint supressions that are no longer needed and
add a new one for a line that pylint is now complaining about.  I'm
reluctant to make changes to the code itself since I don't have time to
dig depper into this so just supresse the complaint for now.