fix: Unpin Pymongo.
We only use very simple objects from pymongo in this repo and none of
them have changes in the intermediate versions.  The only usage of the
SON object is to instantiate one and so no code in this repo is going to
break by this upgrade.

There may be other repos that upstream that might need to pin pymongo
based on their usage but it shouldn't be pinned here.